cu/buffer.h:     cu_buffer_incr_content_end: Bugfix.     cu_buffer_produce: New.
cu/ptr.h: Use pointers to const.

cu/debug.c, cuex/print.c, cuex/semilattice.c: Fix some printf warnings.

+++ cuex/print.c
@@ -133,7 +133,7 @@ cuex_print_ex(cuex_t ex, FILE *out)
 			    cu_debug_unreachable();
 		    }
 		if (cuex_is_ivarmeta(meta))
-		    fprintf(out, "%s%d", prefix, cuex_varmeta_index(meta));
+		    fprintf(out, "%s%ld", prefix, (long)cuex_varmeta_index(meta));
 		else
 		    fprintf(out, "_%s%d", prefix, varindex(ex, 0));
 	    }
